+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/hide/player/flyoutmenupanel/HidePlayerFlyoutMenuPatch.kt
@@ -45,10 +45,10 @@ val hidePlayerFlyoutMenuPatch = bytecodePatch(
                     SwitchPreference("revanced_hide_player_flyout_loop_video"),
                     SwitchPreference("revanced_hide_player_flyout_ambient_mode"),
                     SwitchPreference("revanced_hide_player_flyout_stable_volume"),
+                    SwitchPreference("revanced_hide_player_flyout_listen_with_youtube_music"),
                     SwitchPreference("revanced_hide_player_flyout_help"),
                     SwitchPreference("revanced_hide_player_flyout_speed"),
                     SwitchPreference("revanced_hide_player_flyout_lock_screen"),
-                    SwitchPreference("revanced_hide_player_flyout_more_info"),
                     SwitchPreference(
                         key = "revanced_hide_player_flyout_audio_track",
                         tag = "app.revanced.extension.youtube.settings.preference.HideAudioFlyoutMenuPreference"